Move feeder gateway types to their own files
